backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
azure-ai-projects>=2.0.0b1
azure-identity>=1.15.0
pydantic>=2.0.0
